# haika-api/swagger-spec/api.yaml
openapi: 3.0.1
info:
  title: Haika APIs
  description: 'This is the Rest Api Specification for Haika app'
  termsOfService: ''
  contact:
    email: taras0318@outlook.com
  version: 1.0.0
servers:
  - url: 'http://localhost:3000/api/v1'
...
paths:
...
  '/admin/categories/{user_id}':
    get:
      tags:
        - admin
      summary: Admin view categories
      description: View categories of users by user_id
      operationId: getCategoriesByUserId
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          description: User id that needs to be fetched categores. Use XXXX for testing.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
                  message:
                    type: string
        '400':
          description: Invalid user id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/InvalidToken'        
        '404':
          $ref: '#/components/responses/NotFound' 
        '500':
          $ref: '#/components/responses/ServerError' 
...
  '/admin':
    post:
      tags:
        - admin
      summary: Add a new admin
      description: Add a new admin
      operationId: addAdmin
      requestBody:
        description: Admin object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUser'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: 
                    type: integer
                    format: int32
                  data:
                    type: object
                    properties:
                      token: 
                        type: string
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound' 
        '405':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError' 
      x-codegen-request-body-name: body
...
  '/post/{post_id}':
    put:
      tags:
        - posts
      summary: Update user post
      description: Update post by user id
      operationId: updatePost
      security:
        - bearerAuth: []
      parameters:
        - name: post_id
          in: path
          description: 'Pointed post id. Use XXXX for testing. '
          required: true
          schema:
            type: string    
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: 
                    type: integer
                    format: int32
                  message:
                    type: string
        '400':
          description: Invalid post id
          content: {}
        '401':
          $ref: '#/components/responses/InvalidToken' 
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError' 
  '/post':
    post:
      tags:
        - posts
      summary: Add user post
      description: Add new post 
      operationId: addPost
      security:
        - bearerAuth: []
      requestBody:
        description: Post object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPost'
        required: true    
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: 
                    type: integer
                    format: int32
                  message:
                    type: string         
        '400':
          description: Invalid post id
          content: {}      
        '401':
          $ref: '#/components/responses/InvalidToken' 
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError' 
...
components:
  responses:
    InvalidToken:
      description: Invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    AddUser:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
...
    Category:
      type: object
      properties:
        category_id:
          type: string
        category:
          type: string
        updated_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
...
    AddPost:
      type: object
      properties:
        images:
          type: array
          items: 
            type: object
        list:
          type: array
          items:
            type: string
        category_id: 
          type: string    
        state:
          type: string
        link:
          type: string      
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT